Using libuuid bindings instead of NetworkManager library.
NetworkManager libraries were not available on the build server and it is not a
really good dependency in itself.

Valum depends on uuid >= 2.20 to generate unique session cookie identifier and
distributes its vapi from https://github.com/nemequ/vala-extra-vapis.

mandatory=True is implicit in wscript, so we don't need to specify it.

Adds a simple example of session CRUD.